refactor: move EnglishLevelButton to common (#15)
